backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.1",
    "google-genai>=1.29.0",
    "pydantic>=2.7.0",
    "requests>=2.32.3",
]
